Keep speculative server arg handling local
