+++ python/sglang/srt/speculative/spec_info.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import json
-import logging
 from abc import ABC, abstractmethod
 from enum import Enum, IntEnum, auto
 from typing import TYPE_CHECKING, Callable, List, Optional, Tuple, Type, Union
@@ -25,347 +23,6 @@ if TYPE_CHECKING:
     from sglang.srt.server_args import ServerArgs
     from sglang.srt.speculative.base_spec_worker import BaseSpecWorker
     from sglang.srt.speculative.ngram_worker import NGRAMWorker
-
-
-logger = logging.getLogger(__name__)
-
-
-def _handle_dflash(server_args: ServerArgs) -> None:
-    if server_args.enable_dp_attention:
-        raise ValueError(
-            "Currently DFLASH speculative decoding does not support dp attention."
-        )
-
-    if server_args.pp_size != 1:
-        raise ValueError(
-            "Currently DFLASH speculative decoding only supports pp_size == 1."
-        )
-
-    if server_args.speculative_draft_model_path is None:
-        raise ValueError(
-            "DFLASH speculative decoding requires setting --speculative-draft-model-path."
-        )
-
-    # DFLASH does not use EAGLE-style `num_steps`/`topk`, but those fields still
-    # affect generic scheduler/KV-cache accounting (buffer sizing, KV freeing,
-    # RoPE reservation). Force them to 1 to avoid surprising memory behavior.
-    #
-    # For DFlash, the natural unit is `block_size` (verify window length).
-    if server_args.speculative_num_steps is None:
-        server_args.speculative_num_steps = 1
-    elif int(server_args.speculative_num_steps) != 1:
-        logger.warning(
-            "DFLASH only supports speculative_num_steps == 1; overriding speculative_num_steps=%s to 1.",
-            server_args.speculative_num_steps,
-        )
-        server_args.speculative_num_steps = 1
-
-    if server_args.speculative_eagle_topk is None:
-        server_args.speculative_eagle_topk = 1
-    elif int(server_args.speculative_eagle_topk) != 1:
-        logger.warning(
-            "DFLASH only supports speculative_eagle_topk == 1; overriding speculative_eagle_topk=%s to 1.",
-            server_args.speculative_eagle_topk,
-        )
-        server_args.speculative_eagle_topk = 1
-
-    if server_args.speculative_dflash_block_size is not None:
-        if int(server_args.speculative_dflash_block_size) <= 0:
-            raise ValueError(
-                "DFLASH requires --speculative-dflash-block-size to be positive, "
-                f"got {server_args.speculative_dflash_block_size}."
-            )
-        if server_args.speculative_num_draft_tokens is not None and int(
-            server_args.speculative_num_draft_tokens
-        ) != int(server_args.speculative_dflash_block_size):
-            raise ValueError(
-                "Both --speculative-num-draft-tokens and --speculative-dflash-block-size are set "
-                "but they differ. For DFLASH they must match. "
-                f"speculative_num_draft_tokens={server_args.speculative_num_draft_tokens}, "
-                f"speculative_dflash_block_size={server_args.speculative_dflash_block_size}."
-            )
-        server_args.speculative_num_draft_tokens = int(
-            server_args.speculative_dflash_block_size
-        )
-
-    if server_args.speculative_num_draft_tokens is None:
-        from sglang.srt.speculative.dflash_utils import (
-            parse_dflash_draft_config,
-        )
-
-        model_override_args = json.loads(server_args.json_model_override_args)
-        inferred_block_size = None
-        try:
-            from sglang.srt.utils.hf_transformers_utils import get_config
-
-            draft_hf_config = get_config(
-                server_args.speculative_draft_model_path,
-                trust_remote_code=server_args.trust_remote_code,
-                revision=server_args.speculative_draft_model_revision,
-                model_override_args=model_override_args,
-            )
-            inferred_block_size = parse_dflash_draft_config(
-                draft_hf_config=draft_hf_config
-            ).resolve_block_size(default=None)
-        except Exception as e:
-            logger.warning(
-                "Failed to infer DFLASH block_size from draft model config; "
-                "defaulting speculative_num_draft_tokens to 16. Error: %s",
-                e,
-            )
-
-        if inferred_block_size is None:
-            inferred_block_size = 16
-            logger.warning(
-                "speculative_num_draft_tokens is not set; defaulting to %d for DFLASH.",
-                inferred_block_size,
-            )
-        server_args.speculative_num_draft_tokens = inferred_block_size
-
-    if server_args.speculative_draft_window_size is not None:
-        draft_tokens = int(server_args.speculative_num_draft_tokens)
-        if server_args.speculative_draft_window_size < draft_tokens:
-            raise ValueError(
-                "--speculative-draft-window-size must be >= "
-                "--speculative-num-draft-tokens (block_size). "
-                f"window_size={server_args.speculative_draft_window_size}, block_size={draft_tokens}."
-            )
-
-    if server_args.max_running_requests is None:
-        server_args.max_running_requests = 48
-        logger.warning(
-            "Max running requests is reset to 48 for speculative decoding. You can override this by explicitly setting --max-running-requests."
-        )
-
-    if server_args.enable_mixed_chunk:
-        server_args.enable_mixed_chunk = False
-        logger.warning(
-            "Mixed chunked prefill is disabled because of using dflash speculative decoding."
-        )
-
-
-def _handle_frozen_kv_mtp(server_args: ServerArgs) -> None:
-    if server_args.max_running_requests is None:
-        server_args.max_running_requests = 48
-        logger.warning(
-            "Max running requests is reset to 48 for speculative decoding. You can override this by explicitly setting --max-running-requests."
-        )
-
-    if server_args.enable_mixed_chunk:
-        server_args.enable_mixed_chunk = False
-        logger.warning(
-            "Mixed chunked prefill is disabled because of using "
-            "Frozen-KV MTP speculative decoding."
-        )
-
-
-def _handle_eagle_family(server_args: ServerArgs) -> None:
-    if (
-        server_args.speculative_algorithm == "STANDALONE"
-        and server_args.enable_dp_attention
-    ):
-        # TODO: support dp attention for standalone speculative decoding
-        raise ValueError(
-            "Currently standalone speculative decoding does not support dp attention."
-        )
-
-    if server_args.max_running_requests is None:
-        server_args.max_running_requests = 48
-        logger.warning(
-            "Max running requests is reset to 48 for speculative decoding. You can override this by explicitly setting --max-running-requests."
-        )
-
-    if server_args.disable_overlap_schedule:
-        logger.warning(
-            "Non-overlap (synchronous) spec v2 is used for eagle/eagle3/standalone "
-            "speculative decoding."
-        )
-    else:
-        logger.warning(
-            "Overlap spec v2 is enabled by default for eagle/eagle3/standalone speculative decoding."
-        )
-
-    if server_args.enable_mixed_chunk:
-        server_args.enable_mixed_chunk = False
-        logger.warning(
-            "Mixed chunked prefill is disabled because of using "
-            "eagle speculative decoding."
-        )
-
-    model_arch = server_args.get_model_config().hf_config.architectures[0]
-    if model_arch in [
-        "DeepseekV32ForCausalLM",
-        "DeepseekV3ForCausalLM",
-        "DeepseekV4ForCausalLM",
-        "Glm4MoeForCausalLM",
-        "Glm4MoeLiteForCausalLM",
-        "GlmMoeDsaForCausalLM",
-        "BailingMoeForCausalLM",
-        "BailingMoeV2ForCausalLM",
-        "BailingMoeV2_5ForCausalLM",
-        "MistralLarge3ForCausalLM",
-        "PixtralForConditionalGeneration",
-        "HYV3ForCausalLM",
-    ]:
-        if server_args.speculative_draft_model_path is None:
-            server_args.speculative_draft_model_path = server_args.model_path
-            server_args.speculative_draft_model_revision = server_args.revision
-        else:
-            if model_arch not in [
-                "MistralLarge3ForCausalLM",
-                "PixtralForConditionalGeneration",
-            ]:
-                logger.warning(
-                    "DeepSeek MTP does not require setting speculative_draft_model_path."
-                )
-
-    if (
-        not server_args.speculative_adaptive
-        and server_args.speculative_num_steps is None
-    ):
-        assert (
-            server_args.speculative_eagle_topk is None
-            and server_args.speculative_num_draft_tokens is None
-        )
-
-        (
-            server_args.speculative_num_steps,
-            server_args.speculative_eagle_topk,
-            server_args.speculative_num_draft_tokens,
-        ) = _auto_choose_speculative_params(server_args, model_arch)
-
-    if (
-        server_args.attention_backend == "trtllm_mha"
-        or server_args.decode_attention_backend == "trtllm_mha"
-        or server_args.prefill_attention_backend == "trtllm_mha"
-    ):
-        if server_args.speculative_eagle_topk > 1:
-            raise ValueError(
-                "trtllm_mha backend only supports topk = 1 for speculative decoding."
-            )
-
-    if (
-        server_args.speculative_eagle_topk == 1
-        and server_args.speculative_num_draft_tokens
-        != server_args.speculative_num_steps + 1
-    ):
-        logger.warning(
-            "speculative_num_draft_tokens is adjusted to speculative_num_steps + 1 when speculative_eagle_topk == 1"
-        )
-        server_args.speculative_num_draft_tokens = server_args.speculative_num_steps + 1
-
-    # topk > 1 + page_size > 1 needs the two-pass cascade draft-decode (shared prefix
-    # pass + per-branch expand pass with prefix-tail dup). Only these backends implement
-    # it; flashmla / trtllm_mla / cutlass_mla can't express the per-branch tree, so reject.
-    _PAGE_TREE_SPEC_BACKENDS = ("flashinfer", "fa3", "triton")
-    if (
-        server_args.speculative_eagle_topk > 1
-        and server_args.page_size > 1
-        and server_args.attention_backend not in _PAGE_TREE_SPEC_BACKENDS
-    ):
-        raise ValueError(
-            f"speculative_eagle_topk > 1 with page_size > 1 is only supported on "
-            f"{_PAGE_TREE_SPEC_BACKENDS}; got attention_backend="
-            f"{server_args.attention_backend!r}. Use page_size == 1 or one of those backends."
-        )
-
-
-def _handle_ngram(server_args: ServerArgs) -> None:
-    if not server_args.device.startswith("cuda"):
-        raise ValueError("Ngram speculative decoding only supports CUDA device.")
-
-    if server_args.max_running_requests is None:
-        server_args.max_running_requests = 48
-        logger.warning(
-            "Max running requests is reset to 48 for speculative decoding. You can override this by explicitly setting --max-running-requests."
-        )
-
-    server_args.enable_mixed_chunk = False
-    server_args.speculative_eagle_topk = server_args.speculative_ngram_max_bfs_breadth
-    if server_args.speculative_num_draft_tokens is None:
-        server_args.speculative_num_draft_tokens = 12
-        logger.warning(
-            "speculative_num_draft_tokens is set to 12 by default for ngram speculative decoding. "
-            "You can override this by explicitly setting --speculative-num-draft-tokens."
-        )
-    if server_args.speculative_num_steps is None:
-        server_args.speculative_num_steps = (
-            server_args.speculative_num_draft_tokens
-            // server_args.speculative_eagle_topk
-        )
-    if server_args.speculative_ngram_external_corpus_path is not None:
-        if server_args.speculative_ngram_external_sam_budget <= 0:
-            raise ValueError(
-                "--speculative-ngram-external-sam-budget must be positive when "
-                "--speculative-ngram-external-corpus-path is set."
-            )
-        if server_args.speculative_ngram_external_corpus_max_tokens <= 0:
-            raise ValueError(
-                "--speculative-ngram-external-corpus-max-tokens must be positive when "
-                "--speculative-ngram-external-corpus-path is set."
-            )
-        if (
-            server_args.speculative_ngram_external_sam_budget
-            > server_args.speculative_num_draft_tokens - 1
-        ):
-            raise ValueError(
-                "speculative_ngram_external_sam_budget must be less than or equal to "
-                f"speculative_num_draft_tokens - 1 ({server_args.speculative_num_draft_tokens - 1})."
-            )
-    logger.warning(
-        "The mixed chunked prefill are disabled because of "
-        "using ngram speculative decoding."
-    )
-
-    if (
-        server_args.speculative_eagle_topk > 1
-        and server_args.page_size > 1
-        and server_args.attention_backend != "flashinfer"
-    ):
-        raise ValueError(
-            f"speculative_eagle_topk({server_args.speculative_eagle_topk}) > 1 "
-            f"with page_size({server_args.page_size}) > 1 is unstable "
-            "and produces incorrect results for paged attention backends. "
-            "This combination is only supported for the 'flashinfer' backend."
-        )
-    if server_args.enable_dp_attention:
-        # TODO: support dp attention for ngram speculative decoding
-        raise ValueError(
-            "Currently ngram speculative decoding does not support dp attention."
-        )
-
-
-def _auto_choose_speculative_params(server_args: ServerArgs, model_arch: str) -> tuple:
-    """
-    Automatically choose the parameters for speculative decoding.
-
-    You can tune them on your own models and prompts with scripts/playground/bench_speculative.py
-    """
-    if server_args.speculative_algorithm == "STANDALONE":
-        return (3, 1, 4)
-    if model_arch in ["LlamaForCausalLM"]:
-        return (5, 4, 8)
-    elif model_arch in [
-        "DeepseekV32ForCausalLM",
-        "DeepseekV3ForCausalLM",
-        "DeepseekV2ForCausalLM",
-        "GptOssForCausalLM",
-        "Glm4MoeForCausalLM",
-        "Glm4MoeLiteForCausalLM",
-        "GlmMoeDsaForCausalLM",
-        "BailingMoeForCausalLM",
-        "BailingMoeV2ForCausalLM",
-        "BailingMoeV2_5ForCausalLM",
-        "MistralLarge3ForCausalLM",
-        "PixtralForConditionalGeneration",
-        "MiMoV2ForCausalLM",
-        "MiMoV2FlashForCausalLM",
-    ]:
-        return (3, 1, 4)
-    elif model_arch in ["Grok1ForCausalLM", "Grok1VForCausalLM"]:
-        return (5, 4, 8)
-    else:
-        return (3, 1, 4)
 
 
 class SpeculativeAlgorithm(Enum):
@@ -507,6 +164,13 @@ class SpeculativeAlgorithm(Enum):
 
         In-place updated.
         """
+        from sglang.srt.arg_groups.speculative_hook import (
+            _handle_dflash,
+            _handle_eagle_family,
+            _handle_frozen_kv_mtp,
+            _handle_ngram,
+        )
+
         if self.is_dflash():
             _handle_dflash(server_args)
         elif self.is_frozen_kv_mtp():
